django-polymorphic/requirements.txt
../requirements-base.txt
django-polymorphic==2.1.2